failover graphene subblock to full subblock
new branch that includes primitives for lbck compatibility

introduce merkle subblock (not yet compiling)

small changes to get project to build

small tweak to allow lbck to compile

add MerkleSubBlock Export JNI export

some changes to try to get c++ compiling

+++ src/rpc/tailstorm.cpp
@@ -324,7 +324,7 @@ UniValue getdaginfo(const UniValue &params, bool fHelp)
     }
 
     UniValue obj(UniValue::VOBJ);
-    obj.pushKV("size", tailstormDagSet.Size());
+    obj.pushKV("size", (int) tailstormDagSet.Size());
 
     return obj;
 }
